moving encoding to dataset
